+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
-## v0.2.10 - ??
+## v0.2.10 - 2017-05-19
 ### Added:
  - Search in resume black/white lists
  - Active/archived/hidden vacancies